Added all the changes that I have made to WRF-TOOLS so far (same as the contents of the folder "WRF Tools_Nov_12_2021" in my home directory, see however the last point below).
 * Modified all the files that I read, so that they have the same editing rules
as the ones I use in my general coding (basic deciphering).
 * Added some namelist snippets for ERA-I NA runs.
 * Modified xconfig.sh, according to the requirements for the ERA-I NA runs.
 * Also made minor changes to xconfig.sh to reflect the new setup.
 * Fixed some actual/potential problems in the code.
 * Particularly, depending on -z or -n METDATA and RAMIN being 0 or 1, some scenarios
did not work properly. Some of the changes are meant to fix these.
 * For the case with -n METDATA and RAMIN=0, we may get a few errors due the code
being unable to delete or make files or folders (e.g., if we do not have access to ram),
etc. However, these are minor errors and the code should work overall. These errors were
not fixed at the moment, due to their lower priority.
 * Note that the "WRF Tools_Nov_12_2021" does not have the new fixed wrfout_average.py.